training/requirements.txt
tensorflow>=2.12.0
numpy>=1.23.0
Pillow>=9.0.0
matplotlib>=3.5.0
scikit-learn>=1.2.0
opencv-python-headless>=4.7.0
